# docs/sprint-1/openapi.yaml
openapi: 3.1.0
info:
  title: SmartSweep API
  version: 0.1.0
paths:
...
  /api/v1/wards/me:
    get:
      tags:
      - Wards
      summary: Get My Ward
      operationId: get_my_ward_api_v1_wards_me_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WardRead'
      security:
      - HTTPBearer: []
  /api/v1/wards/{ward_id}:
    get:
      tags:
      - Wards
      summary: Get Ward
      operationId: get_ward_api_v1_wards__ward_id__get
      parameters:
      - name: ward_id
        in: path
        required: true
        schema:
          type: integer
          title: Ward Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WardRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/transparency:
    get:
      tags:
      - Transparency
      summary: List Transparency Posts
      operationId: list_transparency_posts_api_v1_transparency_get
      parameters:
      - name: ward_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ward Id
      - name: page
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          default: 20
          title: Page Size
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Page_TransparencyPostRead_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Transparency
      summary: Create Transparency Post
      operationId: create_transparency_post_api_v1_transparency_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransparencyPostCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransparencyPostRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/transparency/{post_id}:
    get:
      tags:
      - Transparency
      summary: Get Transparency Post
      operationId: get_transparency_post_api_v1_transparency__post_id__get
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: integer
          title: Post Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransparencyPostRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/transparency/{post_id}/applaud:
    post:
      tags:
      - Transparency
      summary: Applaud Transparency Post
      operationId: applaud_transparency_post_api_v1_transparency__post_id__applaud_post
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: integer
          title: Post Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransparencyPostRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/transparency/{post_id}/comments:
    get:
      tags:
      - Transparency
      summary: List Post Comments
      operationId: list_post_comments_api_v1_transparency__post_id__comments_get
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: integer
          title: Post Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PostCommentRead'
                title: Response List Post Comments Api V1 Transparency  Post Id  Comments
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Transparency
      summary: Create Post Comment
      operationId: create_post_comment_api_v1_transparency__post_id__comments_post
      security:
      - HTTPBearer: []
      parameters:
      - name: post_id
        in: path
        required: true
        schema:
          type: integer
          title: Post Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCommentCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostCommentRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    PageMeta:
      properties:
        page:
          type: integer
          minimum: 1.0
          title: Page
          description: 1-based index of the returned page.
        page_size:
          type: integer
          minimum: 1.0
          title: Page Size
          description: Maximum items per page.
        total:
          type: integer
          minimum: 0.0
          title: Total
          description: Total matching rows, ignoring pagination.
        total_pages:
          type: integer
          minimum: 0.0
          title: Total Pages
          description: ceil(total / page_size).
      type: object
      required:
      - page
      - page_size
      - total
      - total_pages
      title: PageMeta
      description: Pagination counters. ``total_pages`` is derived, never supplied.
...
    Page_TransparencyPostRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/TransparencyPostRead'
          type: array
          title: Items
        meta:
          $ref: '#/components/schemas/PageMeta'
      type: object
      required:
      - items
      - meta
      title: Page[TransparencyPostRead]
    PostCommentCreate:
      properties:
        comment:
          type: string
          maxLength: 2000
          minLength: 1
          title: Comment
      type: object
      required:
      - comment
      title: PostCommentCreate
      description: Payload for commenting on a transparency post.
    PostCommentRead:
      properties:
        id:
          type: integer
          title: Id
        post_id:
          type: integer
          title: Post Id
        user_id:
          type: integer
          title: User Id
        comment:
          type: string
          title: Comment
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - post_id
      - user_id
      - comment
      - created_at
      title: PostCommentRead
      description: Post comment response schema.
...
    TransparencyPostCreate:
      properties:
        complaint_id:
          type: integer
          title: Complaint Id
        title:
          type: string
          maxLength: 255
          minLength: 1
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        before_photo_url:
          anyOf:
          - type: string
            maxLength: 2048
          - type: 'null'
          title: Before Photo Url
        after_photo_url:
          anyOf:
          - type: string
            maxLength: 2048
          - type: 'null'
          title: After Photo Url
      type: object
      required:
      - complaint_id
      - title
      title: TransparencyPostCreate
      description: Payload for crew/admin publishing a cleanup record for a resolved
        complaint.
    TransparencyPostRead:
      properties:
        id:
          type: integer
          title: Id
        complaint_id:
          type: integer
          title: Complaint Id
        ward_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Ward Id
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        before_photo_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Before Photo Url
        after_photo_url:
          anyOf:
          - type: string
          - type: 'null'
          title: After Photo Url
        applause_count:
          type: integer
          title: Applause Count
        posted_by_user_id:
          anyOf:
          - type: integer
          - type: 'null'
          title: Posted By User Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - complaint_id
      - title
      - applause_count
      - created_at
      - updated_at
      title: TransparencyPostRead
      description: Transparency post response schema.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WardRead:
      properties:
        name:
          type: string
          maxLength: 255
          minLength: 1
          title: Name
        code:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Code
        zone:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Zone
        description:
          anyOf:
          - type: string
            maxLength: 500
          - type: 'null'
          title: Description
        is_active:
          type: boolean
          title: Is Active
          default: true
        id:
          type: integer
          title: Id
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - name
      - id
      - created_at
      - updated_at
      title: WardRead
      description: Ward response schema.
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer